LayoutTests/fast/canvas/webgl/vertexAttribPointer-with-bad-offset-expected.txt

CONSOLE MESSAGE: line 49: WebGL: INVALID_VALUE: vertexAttribPointer: bad offset
PASS: vertexAttribPointer should have an error.
